chore: run create-plugin@5.27.1 update
