+++ .config/.cprc.json
@@ -1,3 +1,3 @@
 {
-  "version": "5.25.5"
+  "version": "5.27.1"
 }
