+++ package.json
@@ -21,7 +21,7 @@
     "@babel/core": "^7.21.4",
     "@grafana/e2e": "10.4.1",
     "@grafana/e2e-selectors": "10.4.1",
-    "@grafana/eslint-config": "^8.0.0",
+    "@grafana/eslint-config": "^8.2.0",
     "@grafana/tsconfig": "^2.0.0",
     "@swc/core": "^1.3.90",
     "@swc/helpers": "^0.5.0",
@@ -72,15 +72,16 @@
   },
   "dependencies": {
     "@emotion/css": "11.13.5",
-    "@grafana/data": "^12.0.2",
-    "@grafana/runtime": "^12.0.2",
-    "@grafana/schema": "^12.0.2",
-    "@grafana/ui": "^12.0.2",
+    "@grafana/data": "^12.2.0",
+    "@grafana/runtime": "^12.2.0",
+    "@grafana/schema": "^12.2.0",
+    "@grafana/ui": "^12.2.0",
     "graphiql": "^3.1.2",
     "graphql": "^16.8.1",
     "react": "^18.3.1",
     "react-dom": "^18.3.1",
-    "tslib": "2.8.1"
+    "tslib": "2.8.1",
+    "@grafana/i18n": "^12.2.0"
   },
   "packageManager": "npm@8.19.4"
 }
